Split Matrix.jl and MatrixAlgebra.jl and alter documentation to follow.
